requirements.txt
pandas
scikit-learn
joblib
numpy
pytest
boto3
psycopg2-binary
python-dotenv